backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
faiss-cpu
sentence-transformers
